+++ CHANGELOG.md
@@ -34,6 +34,8 @@ we hit release version 1.0.0.
 - A new `AtomicMatrixPlot` to plot sparse matrices, #668
 
 ### Fixed
+- changed `read_force_constant` to `read_hessian`, the old methods are retained with
+  deprecation warnings.
 - `pdosSileSiesta` plotting produced wrong spin components for NC/SOC
 - `tqdm` changed API in 2019, `eta=True` in Notebooks should now work
 - `SparseCSR` ufunc handling, in some corner cases could the dtype casting do things